backend/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pydantic[dotenv]
groq
google-cloud-vision
pdfplumber
transformers
torch
scipy
